+++ CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## [v0.5] - 2024/11/01
+
+### Added
+
+- Tequila
+- Cheese Stout
+
+### Changed
+
+- Recipes:
+  - Sake - added 'minecraft:water_bucket' liquid requirement.
+  - Cahors - added 'italian_delight:sweet_berry_wine' liquid requirement.
+- Beverages:
+  - Braga's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Braga's effect set to strength for 2 minutes, speed 2 for 3 minutes.
+  - Brandy's tipsy effect strength set to 2, duration set to 4 minutes.
+  - Brandy's effect set to water breathing for 3 minutes.
+  - Cider's tipsy effect strength set to 2, duration set to 3 minutes.
+  - Cider's effect set to regeneration 2 for 3 minutes.
+  - Gin's tipsy effect strength set to 3, duration set to 1 minute.
+  - Gin's effect set to speed 3 for 4 minutes.
+  - Kvass' tipsy effect strength set to 0.
+  - Kvass' effect set to instant health.
+  - Kvass' nutrition set to 3, saturation to 1.
+  - Liqueur's tipsy effect strength set to 3, duration set to 4 minutes.
+  - Liqueur's effect set to regeneration 2 for 4 minutes.
+  - Martini's tipsy effect strength set to 2, duration set to 2 minutes.
+  - Martini's effect set to comfort for 2 minutes.
+  - Melon schnapps' tipsy effect strength set to 2, duration set to 15 minutes.
+  - Melon schnapps' effect set to blindness for 3 minutes, strength 2 for 3 minutes.
+  - Sake's tipsy effect strength set to 1, duration set to 4 minutes.
+  - Sake's effect set to absorption 3 for 8 minutes.
+  - Sbiten's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Sbiten's effect set to haste 2 for 8 minutes.
+  - Sbiten's nutrition set to 6, saturation to 3.
+  - Whisky's tipsy effect strength set to 3, duration set to 19 minutes.
+  - Whisky's effect set to strength 2 for 7 minutes.
+  - Bourbon's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Bourbon's effect set to fire resistance for 8 minutes.
+  - Corn whisky's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Corn whisky's effect set to levitation 2 for 1 minute.
+  - Moonshine's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Moonshine's effect set to health boost 3 for 5 minutes.
+  - Cahor's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Cahor's effect set to invisibility for 2 minutes, regeneration 3 for 2 minutes.
+  - Champagne's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Champagne's effect set to jump boost 2 for 3 minutes.
+  - Cognac's tipsy effect strength set to 1, duration set to 2 minutes.
+  - Cognac's effect set to resistance 3 for 4 minutes.
+
+### Fixed
+
+- Beverages not giving appropriate potion effects after drinking.
+- Beverages stacking up to 64 instead of 16.
+- Beverages not giving empty tankard back after drinking.
+
 ## [v0.4] - 2024/10/28
 
 ### Changed